Update ArtistInfoRepository and create LastFMLocalStorage
