+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/ArtistInfoRepository.kt
@@ -1,35 +1,22 @@
 package ayds.lisboa.songinfo.moredetails.fulllogic
 
 import android.content.ContentValues
-import android.content.Context
-import android.database.Cursor
-import android.database.sqlite.SQLiteDatabase
-import android.database.sqlite.SQLiteOpenHelper
 import ayds.lisboa.songinfo.moredetails.fulllogic.ArtistInfo.LastFmArtistInfo
-import java.sql.SQLException
 
 interface ArtistInfoRepository{
     fun saveArtist(artist: String, artistInfo: LastFmArtistInfo)
     fun getArtistInfo(artist: String): LastFmArtistInfo?
 }
 
-internal class ArtistInfoRepositoryImpl(context: Context?) : SQLiteOpenHelper(context, DB_NAME, null, 1), ArtistInfoRepository {
-    private val projection = arrayOf(
-        ID,
-        ARTIST,
-        BIO_CONTENT,
-        URL
-    )
+internal class ArtistInfoRepositoryImpl(
+    private val lastFMLocalStorage: LastFMLocalStorage,
+    private val lastFMService: LastFMService
+) : ArtistInfoRepository {
 
-    override fun onCreate(db: SQLiteDatabase) {
-        db.execSQL(CREATE_ARTIST_INFO_TABLE)
-    }
-
-    override fun onUpgrade(db: SQLiteDatabase, oldVersion: Int, newVersion: Int) {}
 
     override fun saveArtist(artist: String, artistInfo: LastFmArtistInfo) {
         val artistValues = getArtistValues(artist, artistInfo)
-        this.writableDatabase.insert(TABLE_ARTISTS, null, artistValues)
+        lastFMLocalStorage.saveArtist(artistValues)
     }
 
     private fun getArtistValues(artist: String, artistInfo: LastFmArtistInfo): ContentValues {
@@ -46,24 +33,10 @@ internal class ArtistInfoRepositoryImpl(context: Context?) : SQLiteOpenHelper(co
     override fun getArtistInfo(artist: String): LastFmArtistInfo? {
         val artistValues = arrayOf(artist)
 
-        return cursorHandling(artistValues)
+        return lastFMLocalStorage.getArtistInfo(artistValues)
     }
 
-    private fun cursorHandling(artistValues: Array<String>): LastFmArtistInfo? {
-        val cursor = this.readableDatabase.query(
-            TABLE_ARTISTS,
-            projection,
-            ARTIST_COLUMN,
-            artistValues,
-            null,
-            null,
-            RESULT_SET_ORDER
-        )
-        val artistInfo = map(cursor)
-        cursor.close()
 
-        return artistInfo
-    }
 
 
 }
